+++ Source/PluginEditor.h
@@ -15,6 +15,7 @@
 #include "UI/PatternVisualizer.h"
 #include "UI/PresetBrowser.h"
 #include "UI/PolyrhythmLayerEditor.h"
+#include "UI/MidiActivityPane.h"
 
 class GenerativeMIDIEditor : public juce::AudioProcessorEditor,
                               private juce::Timer,
@@ -58,6 +59,7 @@ private:
     // Pattern area components
     PatternVisualizer patternDisplay;
     std::unique_ptr<PolyrhythmLayerEditor> polyLayerEditor;
+    MidiActivityPane midiActivityPane;
 
     // Parameter sliders
     juce::Slider tempoSlider;